Replace use of ArrangeBlock with generic Block
